Ask for assistance
The men are close to rebellion. We have seen no trace of fresh water
for three days, and our supplies are almost out. The natives have not
yet attacked, but occasionally we hear rustling in the bushes and light
glints from rough-cut spear heads. The monstrous lizards that dwell on
these isles will not leave us alone, but we dare not waste our precious
gunpowder killing them; more always appear. It will not be long now.

Adding a Flattr button below download link in the hope that help comes.

[This is the last entry in the diary.]